dev: refactor: merge Text.Megaparsec.Custom into Hledger.Utils.Parse
